requirements.txt
# Core dependencies
pandas==2.2.1
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.1

# Reddit API
praw==7.7.1

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0

# Data processing
beautifulsoup4==4.12.2
lxml==5.1.0

# Date handling
python-dateutil==2.8.2

# Logging
colorlog==6.7.0

# Progress tracking
tqdm==4.66.1

# Optional: Discord API (if implementing direct bot)
# discord.py==2.3.2

# Optional: Advanced analysis
# scipy==1.12.0
# nltk==3.8.1
# scikit-learn==1.3.2